Implement MP-synchronzed microtime(), based on Jason Thorpe's port of Dave Mills's cycle-counter-based microtime.
XXX We need to figure out how to deal with processors where the cycle
counter speed varies.  We don't have a whole lot of alternatives; we
could only use this on multiprocessors, which are less likely to
throttle the clock to save power.

The local apic timer runs at a fixed frequency, but it's
count-down-to-zero, not the count-up-forever type of timer which the
Mills code expects.

This also does not solve the problem of microtime on 486-based
multiprocessors..