Pull up following revision(s) (requested by bouyer in ticket #678): 	distrib/sets/lists/man/mi: revision 1.1001 	doc/CHANGES: revision 1.853 via patch 	share/man/man4/Makefile: revision 1.428 	share/man/man4/jmide.4: revision 1.1 	sys/arch/amd64/conf/GENERIC: revision 1.143 	sys/arch/amd64/conf/INSTALL: revision 1.69 	sys/arch/i386/conf/GENERIC: revision 1.827 	sys/arch/i386/conf/INSTALL_LARGE: revision 1.2 via patch 	sys/arch/i386/conf/XEN2_DOM0: revision 1.26 	sys/conf/files: revision 1.844 via patch 	sys/dev/ic/ahcisata_core.c: revision 1.1 via patch 	sys/dev/ic/ahcisata_core.c: revision 1.3 	sys/dev/ic/ahcisatareg.h: revision 1.1 	sys/dev/ic/ahcisatavar.h: revision 1.1 	sys/dev/pci/ahcisata.c: delete 	sys/dev/pci/ahcisatareg.h: delete 	sys/dev/pci/ahcisata_pci.c: revision 1.1 	sys/dev/pci/files.pci: revision 1.287-1.288 	sys/dev/pci/jmide.c: revision 1.1-1.2 	sys/dev/pci/jmide_reg.h: revision 1.1 	sys/dev/pci/pcidevs: revision 1.878 Add JMicron Technology vendor ID, and their current PCIe SATA/PATA controllers.
Split the ahcisata driver in pci front-end and bus-independant back-end.
add jmide(4), a driver for the JMicron Technology JMB36x PCIe to SATA II/PATA
controllers. These controllers can be found on add-on PCIe cards, or
on some motherboards to provide the PATA connectivity (e.g. some intel
ICH8-based motherboards).
Thanks to JMicron Technology for providing me documentation and
different sample boards for this work.

Move mapping of AHCI register so that jmide knows if it fails, and avoids
calling ahci_intr() (which would cause a panic).
Try to use the pciide function for SATA drives if attaching ahci fails (this
doesn't seems to work though, it may be BIOS dependant).
Thanks to Gary Duzan for testing multiple pacthes.

We're not ready to handle ATAPI yet so just claim there's no drive to the
upper layer. This should work around a NULL pointer dereference when
an ATAPI device is detected on a AHCI device.