Rewrite the ACPI Embedded Controller handler to use pure event driven operation. On suspend a special flag is set to force explicit polling as AcpiLeaveSleep accesses the interrupt before GPE handling is restored. The driver uses a kernel thread to handle GPE queries and mutex/condvar for synchronisation.
Split the ACPI EC into two drivers, one that is attached directly by
acpi.c for the ECDT driven attachment and the normal acpiec for late
attachment. Share almost all code between this two drivers. If acpiecdt
is attached, acpiec is returning. This was discussed with cube@ and is
the best solution so far.