Add code to adjust how we generate upcalls in sa_yield(). The key problem is that we can only return from sys_sa_yield() iff there are upcalls, and in some cases we can only tell that there are upcalls to deliver after we have run userret().
So if we get woken up in a way that might not have generated upcalls
(timers, signals), try delivering upcalls. If we delivered something,
exit.

Add a flag to the vp & set it if we do this. In userret(), clear
the flag if set. Also, clear it in sa_upcall_userret() iff we get
into there.

Add comments describing the above so that the next person looking at
the code can understand it.