Pull up following revision(s) (requested by jmmv in ticket #1563): 	dist/atf/tests/atf/tools/t_atf_config.sh        patch 	lib/libatf-c/Makefile                           patch
Fix the values of atf_arch and atf_machine: the former is supposed to be
the cpu name and the latter the port name.  They had been reversed until
now due to some "smart" stupidity^Wlogic in the upstream configure script,
which is now gone.
This is a pullup of revision f9329ca68da7e8557e0803b5747a12f8c10b1258
plus the corresponding reachover build changes.
Addresses PR bin/44305.