Add test cases for PR bin/6764: 'trap ... 0' is supposed to execute the command when the shell exits but it does not work in ksh when the shells exits "implicitly" (without an explicit exit/return statement).
These new tests cover both sh and ksh.  The ksh part of this does not
strictly belong to this directory, but I think it'd be nice to extend
all the tests in here to cover both interpreters whenever that makes
sense, much like we do with the file system tests.