Add placeholder test programs for share/mk files.
The main goal of this change is to simplify the further addition of test
cases for the share/mk infrastructure by adding a few placeholder test
programs to the tests tree.

To not leave these test programs empty, I have added a bunch of extremely
simple test cases to them.