Re-add a line from NetBSD 0.8 to FreeBSD 1.0, which got lost in the last commit.  From the original commit log of rev 1.33:
revision 1.33
date: 2006/03/02 16:21:53;  author: tls;  state: Exp;  lines: +3 -3
Add line from NetBSD 0.8 to FreeBSD 1.0, reflecting the fact that the
patchkit maintainers were provided with a prerelease snapshot of NetBSD,
changes from which they integrated into their working tree, well prior
to their announcement of the FreeBSD project.  Per personal communications
with Charles Hannum and Chris Demetriou.  This line is in at least one
other published version of this chart, but seems to never have been in
the version in our repository (unfortunately, cvs diff on ascii art is
not very helpful).