Remove stray KERNEL_UNLOCK_ONE() in error path of lfs_markv(). From Wolfgang Stukenbrock in PR 44370.
This error path is only reachable if lfs_markv is handed an out of
range inode number, so it's unlikely that it gets tickled very often.

It isn't clear to me that we need the kernel lock in here at all, as
the path to lfs_markv that's actually used at this point (via fcntl)
doesn't take it. But, one thing at a time.