Apply patch, requested by msaitoh in ticket #983: 	sys/dev/mii/brgphy.c			1.68 via patch 	sys/dev/pci/if_bge.c			1.260-1.262 via patch 	sys/dev/pci/if_bgereg.h			1.85-1.86 via patch 	sys/dev/pci/if_bgevar.h			1.17 via patch
Set maximum read byte count to 2048 for PCI-X BCM5703/5704 devices.
For PCI-X BCM5704, set maximum outstanding split transactions to 0.
Same as Linux tg3 and FreeBSD (part of FreeBSD r204978).
Add 40bit DMA bug workaround(BGEF_40BIT_BUG) from FreeBSD
(part of r199670). This workaround is for 5714/5715 controllers and
is not actually a MAC controller bug but an issue with the embedded
PCIe to PCI-X bridge in the device.  This change
uses bus_dmatag_subregion(), so this workaround won't work on some
archs which doesn't support bus_dmatag_subregion().