Use offsets to access the TPR, and not lapic_tpr.
For the record: I'm not sure exactly why it was originally done this way,
and the cvs logs are not quite enlightening. It might have been a way to
speed up the asm access, but it would sound a little bizarre since there
are many more legitimate reasons to do it on the EOI register instead.